+++ Samples/Markup/ImageMarkupProcessor.cs
@@ -40,7 +40,7 @@ public partial class ImageMarkupProcessor : ReplacementMarkupHandler
         LineProvider.RegisterMarkerProcessor("img", this);
     }
 
-    public override List<LineParser.MarkupDiagnostic> ProcessReplacementMarker(MarkupAttribute marker,
+    public override ReplacementMarkerResult ProcessReplacementMarker(MarkupAttribute marker,
         StringBuilder childBuilder, List<MarkupAttribute> childAttributes,
         string localeCode)
     {
@@ -48,14 +48,18 @@ public partial class ImageMarkupProcessor : ReplacementMarkupHandler
         {
             // replace with <image tag> 
             childBuilder.Insert(0, value);
-            return [];
+            return new ReplacementMarkerResult();
         }
 
         if (marker.Name == "img")
         {
             if (!marker.TryGetProperty("path", out MarkupValue imagePath))
             {
-                return [new("No path attribute specified for img markup tag.")];
+                var error = new List<LineParser.MarkupDiagnostic>
+                {
+                    new LineParser.MarkupDiagnostic("No path attribute specified for img markup tag.")
+                };
+                return new ReplacementMarkerResult(error, 0);
             }
 
             var widthString = "";
@@ -75,9 +79,8 @@ public partial class ImageMarkupProcessor : ReplacementMarkupHandler
             // generic image markup
             childBuilder.Insert(0, $"[img{argsString}]res://Samples/Markup/images/{imagePath.StringValue}[/img]");
 
-            return NoDiagnostics;
         }
 
-        return [];
+        return new ReplacementMarkerResult();
     }
 }