+++ addons/YarnSpinner-Godot/Runtime/CustomMarker.cs
@@ -11,5 +11,6 @@ public partial class CustomMarker : Resource
     [Export] public string? Marker;
     [Export] public string? Start;
     [Export] public string? End;
-    [Export] public int MarkerOffset = 0;
+    [Export] public int MarkerOffset;
+    [Export] public int TotalVisibleCharacterCount;
 }