+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.3.11] 2025-xx-xx
+## [0.3.12] 2025-12-x
+
+* Update the base YarnSpinner libraries to 3.1.0. See details here: https://github.com/YarnSpinnerTool/YarnSpinner/releases/tag/v3.1.0
+* Removed the `ReplacementMarkupHandler.NoDiagnostics` static property, as this no longer matches any need due to core changes around replacement markup.
+* `ReplacementMarkuphandler` now returns a `ReplacementMarkerResult` instead of a list of diagnostics.
+* This is to match the behaviour change in core to fix a markup offset bug
+
+## [0.3.11] 2025-12-01
 
 * Port the solution to [this issue](https://github.com/YarnSpinnerTool/IssuesDiscussion/issues/32), making StartDialogue and Stop on the DialogueRunner async methods.
 * Add ChangeDialogue(string nodeName) method to the dialogue runner to allow stopping the dialogue,